+++ CHANGES.rst
@@ -33,7 +33,7 @@ Imviz
 
 - Catalog Search now shows a table of selected entries and allows selecting/deselecting via a tool in the image viewer. [#3429]
 
-- Virtual Observatory plugin to query resources and download data products. [#2872]
+- Virtual Observatory plugin to query resources and download data products. [#2872, #3470]
 
 Mosviz
 ^^^^^^